User mgt

+++ src/main/java/com/volunteer/main/repositories/RoleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends CrudRepository<RoleEntity, Integer> {
+public interface RoleRepository extends CrudRepository<RoleEntity, Long> {
 
 }
