+++ src/main/java/com/volunteer/main/repositories/VolunteerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VolunteerRepository extends CrudRepository<VolunteerEntity, Integer> {
+public interface VolunteerRepository extends CrudRepository<VolunteerEntity, Long> {
 
 }
